feat: SessionManager 분리, Host와 Peer 구분 로직 구현(#7)
